ランダムルール生成器を追加
